+++ doc/source/changelog.rst
@@ -5,6 +5,7 @@ Change Log
 v1.2.0
 ======
 
+* Implemented `Allow callback and initialize to be aliases of each other. <https://github.com/bckohan/django-typer/issues/66>`_
 * Implemented `Shell completion for --pythonpath <https://github.com/bckohan/django-typer/issues/65>`_
 * Implemented `Shell completion for --settings <https://github.com/bckohan/django-typer/issues/64>`_
 * Fixed `An intelligible exception should be thrown when a command is invoked that has no implementation. <https://github.com/bckohan/django-typer/issues/63>`_